ai-worker/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
javalang==0.13.0
sentence-transformers==2.7.0
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic==2.7.0
python-dotenv==1.0.1
